--- a/src/compile-dts.ts
+++ b/src/compile-dts.ts
@@ -3,7 +3,6 @@ import * as ts from 'typescript';
 import { verboseLog, warnLog } from './logger';
 
 import { getCompilerOptions } from './get-compiler-options';
-import { getAbsolutePath } from './helpers/get-absolute-path';
 import { checkProgramDiagnosticsErrors, checkDiagnosticsErrors } from './helpers/check-diagnostics-errors';
 
 export interface CompileDtsResult {
@@ -49,19 +48,21 @@ export function compileDts(rootFiles: readonly string[], preferredConfigPath?: s
 		compilerOptions.composite = undefined;
 	}
 
-	const dtsFiles = getDeclarationFiles(rootFiles, compilerOptions);
+	const host = createCachingCompilerHost(compilerOptions);
 
-	verboseLog(`dts cache:\n  ${Object.keys(dtsFiles).join('\n  ')}\n`);
-
-	const host = ts.createCompilerHost(compilerOptions);
+	const dtsFiles = getDeclarationFiles(rootFiles, compilerOptions, host);
 
 	if (!followSymlinks) {
+		// note that this shouldn't affect the previous call as there we actually want to use actual path in order to compile files
+		// and avoid issues like "you have .ts files in node_modules"
 		host.realpath = (p: string) => p;
 	}
 
+	const moduleResolutionCache = ts.createModuleResolutionCache(host.getCurrentDirectory(), host.getCanonicalFileName, compilerOptions);
+
 	host.resolveModuleNameLiterals = (moduleLiterals: readonly ts.StringLiteralLike[], containingFile: string): ts.ResolvedModuleWithFailedLookupLocations[] => {
 		return moduleLiterals.map((moduleLiteral: ts.StringLiteralLike): ts.ResolvedModuleWithFailedLookupLocations => {
-			const resolvedModule = ts.resolveModuleName(moduleLiteral.text, containingFile, compilerOptions, host).resolvedModule;
+			const resolvedModule = ts.resolveModuleName(moduleLiteral.text, containingFile, compilerOptions, host, moduleResolutionCache).resolvedModule;
 			if (resolvedModule && !resolvedModule.isExternalLibraryImport) {
 				const newExt = declarationExtsRemapping[resolvedModule.extension];
 
@@ -80,14 +81,11 @@ export function compileDts(rootFiles: readonly string[], preferredConfigPath?: s
 
 	const originalGetSourceFile = host.getSourceFile;
 	host.getSourceFile = (fileName: string, languageVersion: ts.ScriptTarget, onError?: (message: string) => void) => {
-		const absolutePath = getAbsolutePath(fileName);
-		const storedValue = dtsFiles.get(absolutePath);
+		const storedValue = dtsFiles.get(host.getCanonicalFileName(fileName));
 		if (storedValue !== undefined) {
-			verboseLog(`dts cache match: ${absolutePath}`);
 			return ts.createSourceFile(fileName, storedValue, languageVersion);
 		}
 
-		verboseLog(`dts cache mismatch: ${absolutePath} (${fileName})`);
 		return originalGetSourceFile(fileName, languageVersion, onError);
 	};
 
@@ -103,6 +101,26 @@ export function compileDts(rootFiles: readonly string[], preferredConfigPath?: s
 	warnAboutTypeScriptFilesInProgram(program);
 
 	return { program, rootFilesRemapping };
+}
+
+function createCachingCompilerHost(compilerOptions: ts.CompilerOptions): ts.CompilerHost {
+	const host = ts.createIncrementalCompilerHost(compilerOptions);
+
+	const sourceFilesCache = new Map<string, ts.SourceFile | undefined>();
+
+	const originalGetSourceFile = host.getSourceFile;
+	host.getSourceFile = (fileName: string, languageVersion: ts.ScriptTarget, onError?: (message: string) => void): ts.SourceFile | undefined => {
+		const key = host.getCanonicalFileName(fileName);
+		let cacheValue = sourceFilesCache.get(key);
+		if (cacheValue === undefined) {
+			cacheValue = originalGetSourceFile(fileName, languageVersion, onError);
+			sourceFilesCache.set(key, cacheValue);
+		}
+
+		return cacheValue;
+	};
+
+	return host;
 }
 
 function changeExtensionToDts(fileName: string): string {
@@ -133,7 +151,7 @@ function changeExtensionToDts(fileName: string): string {
 /**
  * @description Compiles source files into d.ts files and returns map of absolute path to file content
  */
-function getDeclarationFiles(rootFiles: readonly string[], compilerOptions: ts.CompilerOptions): Map<string, string> {
+function getDeclarationFiles(rootFiles: readonly string[], compilerOptions: ts.CompilerOptions, host: ts.CompilerHost): Map<string, string> {
 	// we must pass `declaration: true` and `noEmit: false` if we want to generate declaration files
 	// see https://github.com/microsoft/TypeScript/issues/24002#issuecomment-550549393
 	// also, we don't want to generate anything apart from declarations so that's why `emitDeclarationOnly: true` is here
@@ -146,7 +164,11 @@ function getDeclarationFiles(rootFiles: readonly string[], compilerOptions: ts.C
 		emitDeclarationOnly: true,
 	};
 
-	const program = ts.createProgram(rootFiles, compilerOptions);
+	// theoretically this could be dangerous because the compiler host is created with compiler options
+	// so technically `compilerOptions` and ones that were used to create the host might be different (and most likely will be)
+	// but apparently a compiler host doesn't use compiler options that much, just a few encoding/newLine oriented
+	// so hopefully it should be fine
+	const program = ts.createProgram(rootFiles, compilerOptions, host);
 	const allFilesAreDeclarations = program.getSourceFiles().every((s: ts.SourceFile) => s.isDeclarationFile);
 
 	const declarations = new Map<string, string>();
@@ -161,7 +183,7 @@ function getDeclarationFiles(rootFiles: readonly string[], compilerOptions: ts.C
 
 	const emitResult = program.emit(
 		undefined,
-		(fileName: string, data: string) => declarations.set(getAbsolutePath(fileName), data),
+		(fileName: string, data: string) => declarations.set(host.getCanonicalFileName(fileName), data),
 		undefined,
 		true
 	);