Fixes #302

With the following input _declaration_ file:

```ts
// index.d.ts
import { Context } from 'effect';

export declare const foobar: <Self, Shape>() => Context.TagClass<Self, string, Shape>;
export type A = string;

export {};
```

The execution time of `./node_modules/.bin/dts-bundle-generator src/index.d.ts --no-check` dropped from 2.64s to 1.26s (more than 2x), which is much closed to just running `tsc` 0.96s (but still not ideal).

| cmd | NPM | PR | Diff |
|---|---|---|---|
| `dts-bundle-generator src/index.d.ts --no-check` [^1] | 2.64s | 1.26s | -1.38s (-52%) |
| `dts-bundle-generator src/index.d.ts` [^2] | 3.45s | 2.07s | -1.38s (-40%) |
| `dts-bundle-generator src/index.ts --no-check` | 3.3s | 1.8s | -1.5s (-45%) |
| `dts-bundle-generator src/index.ts` [^2] | 4s | 2.52s | -1.48s (-37%) |

[^1]: If all input files are declaration files then it goes much faster because there is no need in extra
[^2]: `--no-check` disables extra compilation after the job is done thus is faster, but less safer

So overall it is around 40-50% execution time decrease on this particular example. It is worth to say that `effect` is a massive library and it pushes the compiler a lot in terms of types, parsing/processing/validation time.