@timocov 2024-02-20
As the next step I'll try to reduce validation time as technically we can re-use the same caches during the validation time (when running without `--no-check`) to avoid processing the same files again and again, probably get drop a few extra hundreds of ms.